fixed issue with new score popup-message, added gif & collision sound
