--- Snake/.gitignore
@@ -1,3 +1,0 @@
-.snakevenv
-temp_main.py
-test.py
