--- Snake/best_scores.json
@@ -1,1 +1,0 @@
-[]
