+++ assets/sounds/collision_self.wav
