+++ best_scores.json
@@ -1,20 +1,8 @@
 [
     {
-        "user":"Snake",
-        "score":3,
-        "time":"00:13",
-        "date":"2024-12-08 23:19"
-    },
-    {
-        "user":"Test1",
+        "user":"Jo",
         "score":1,
-        "time":"00:03",
-        "date":"2024-12-08 23:17"
-    },
-    {
-        "user":"Test3",
-        "score":0,
-        "time":"00:01",
-        "date":"2024-12-08 23:33"
+        "time":"00:06",
+        "date":"2024-12-14 13:01"
     }
 ]
